requirements.txt
fastapi[standard]
uvicorn
redis
boto3
python-dotenv